Renamed functions as per code review request
