+++ src/domain/transferParticipantStateChange/index.js
@@ -24,12 +24,12 @@
  ******/
 
 const ErrorHandler = require('@mojaloop/central-services-error-handling')
-const TransferFulfilModel = require('../../models/transferFulfil')
+const TransferParticipantStateChangeModel = require('../../models/transferParticipantStateChange')
 
 module.exports = {
   processMsgFulfil: async function (transferEventId, transferEventStateStatus) {
     try {
-      await TransferFulfilModel.updateStateChange(transferEventId, transferEventStateStatus)
+      await TransferParticipantStateChangeModel.updateStateChange(transferEventId, transferEventStateStatus)
       return true
     } catch (err) {
       throw ErrorHandler.Factory.reformatFSPIOPError(err)
