+++ src/handlers/transferParticipantStateChange/handler.js
@@ -23,7 +23,7 @@
  - Name Surname <name.surname@gatesfoundation.com>
 
  * ModusBox
- - Georgi Georgiev <georgi.georgiev@modusbox.com>
+ - Deon Botha <deon.botha@modusbox.com>
  --------------
  ******/
 'use strict'
@@ -40,7 +40,7 @@ const Kafka = require('@mojaloop/central-services-shared').Util.Kafka
 const Logger = require('@mojaloop/central-services-logger')
 const Producer = require('@mojaloop/central-services-stream').Util.Producer
 const retry = require('async-retry')
-const transferFulfilService = require('../../domain/transferFulfil')
+const transferParticipantStateChangeService = require('../../domain/transferParticipantStateChange')
 const Utility = require('@mojaloop/central-services-shared').Util
 const location = { module: 'TransferFulfilHandler', method: '', path: '' } // var object used as pointer
 const consumerCommit = true
@@ -53,7 +53,7 @@ const retryOpts = {
   maxTimeout: retryDelay
 }
 
-const processTransferFulfil = async (error, messages) => {
+const processTransferParticipantStateChange = async (error, messages) => {
   if (error) {
     Logger.error(error)
     throw ErrorHandling.Factory.reformatFSPIOPError(error)
@@ -61,7 +61,7 @@ const processTransferFulfil = async (error, messages) => {
   Logger.info(Utility.breadcrumb(location, messages))
   let message = {}
   try {
-    Logger.info(Utility.breadcrumb(location, { method: 'processTransferFulfill' }))
+    Logger.info(Utility.breadcrumb(location, { method: 'processTransferParticipantStateChange' }))
     if (Array.isArray(messages)) {
       message = messages[0]
     } else {
@@ -80,7 +80,7 @@ const processTransferFulfil = async (error, messages) => {
 
     if (!payload) {
       Logger.info(Utility.breadcrumb(location, `missingPayload--${actionLetter}1`))
-      const fspiopError = ErrorHandling.Factory.createInternalServerFSPIOPError('Transfer fulfil handler missing payload')
+      const fspiopError = ErrorHandling.Factory.createInternalServerFSPIOPError('TransferParticipantStateChange handler missing payload')
       const eventDetail = { functionality: Enum.Events.Event.Type.NOTIFICATION, action: Enum.Events.Event.Action.SETTLEMENT_WINDOW }
       await Kafka.proceed(Config.KAFKA_CONFIG, params, { consumerCommit, fspiopError: fspiopError.toApiErrorObject(Config.ERROR_HANDLING), eventDetail, fromSwitch })
       throw fspiopError
@@ -89,7 +89,7 @@ const processTransferFulfil = async (error, messages) => {
 
     if (transferEventAction === Enum.Events.Event.Action.COMMIT || transferEventAction === Enum.Events.Event.Action.ABORT) {
       await retry(async () => { // use bail(new Error('to break before max retries'))
-        await transferFulfilService.processMsgFulfil(transferEventId, transferEventStateStatus)
+        await transferParticipantStateChangeService.processMsgFulfil(transferEventId, transferEventStateStatus)
         Logger.info(Utility.breadcrumb(location, `done--${actionLetter}2`))
         return true
       }, retryOpts)
@@ -109,10 +109,10 @@ const processTransferFulfil = async (error, messages) => {
  * Calls createHandler to register the handler against the Stream Processing API
  * @returns {boolean} - Returns a boolean: true if successful, or throws and error if failed
  */
-const registerTransferFulfillHandler = async () => {
+const registerTransferParticipantStateChange = async () => {
   try {
     const transferFulfillHandler = {
-      command: processTransferFulfil,
+      command: processTransferParticipantStateChange,
       topicName: Kafka.transformGeneralTopicName(Config.KAFKA_CONFIG.TOPIC_TEMPLATES.GENERAL_TOPIC_TEMPLATE.TEMPLATE, Enum.Events.Event.Type.NOTIFICATION, Enum.Events.Event.Action.EVENT),
       config: Kafka.getKafkaConfig(Config.KAFKA_CONFIG, Enum.Kafka.Config.CONSUMER, Enum.Events.Event.Type.NOTIFICATION.toUpperCase(), Enum.Events.Event.Action.EVENT.toUpperCase())
     }
@@ -135,7 +135,7 @@ const registerTransferFulfillHandler = async () => {
  */
 const registerAllHandlers = async () => {
   try {
-    await registerTransferFulfillHandler()
+    await registerTransferParticipantStateChange()
     return true
   } catch (err) {
     throw ErrorHandling.Factory.reformatFSPIOPError(err)
@@ -143,7 +143,7 @@ const registerAllHandlers = async () => {
 }
 
 module.exports = {
-  processTransferFulfil,
+  processTransferParticipantStateChange,
   registerAllHandlers,
-  registerTransferFulfillHandler
+  registerTransferParticipantStateChange
 }
